requirements.txt
pandas
numpy
xgboost
scikit-learn
joblib
streamlit
plotly
requests
python-dotenv
kagglehub

